pyproject.toml
[project]
name = "pyprize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.3",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.21",
]

[tool.uv]
dev-dependencies = [
    "pip-audit>=2.7.3",
    "ruff>=0.6.3",
    "uvicorn>=0.30.6",
]
